Updated folder layout.
Substitutes are now specifically designated as such in the database.

Individual students are rejected for time conflicts and the rest are saved to the database. The rejected students are individually noted along with what times were rejected for the coordinate to view.

Staff attendance history added.

Easier session/staff/student page navigation as Rod suggested.

Removed Spring 2023, Summer Two 2022, and set Summer One and Fall as Summer and Academic Year respectively.